Move argparsing to a separate module
